Backend/backend_fastapi/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
pymysql==1.1.1
bcrypt==3.2.2
prometheus-client==0.20.0
sentry-sdk==1.45.0
httpx==0.27.0

